pintor-translator_app/src/requirements.txt
altgraph==0.17.3
certifi==2023.5.7
chardet==3.0.4
click==8.1.3
colorama==0.4.6
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.2.0
langdetect==1.0.9
nltk==3.8.1
pefile==2023.2.7
Pillow==9.5.0
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pywin32-ctypes==0.2.0
regex==2023.5.5
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
textblob==0.17.1
tqdm==4.65.0
